+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* Experimental support for non-prefix completion.
 * Starting with Emacs version 24.4, `company-capf` is included in
   `company-backends` and replaces `company-elisp`.
 * `company-capf` supports completion tables that return non-default boundaries.